refactor: move sub-components to components folder.
