+++ packages/ui/FieldBox/components/index.ts
@@ -1,0 +1,2 @@
+export * from './Label';
+export * from './BottomAddon';
