+++ packages/ui/FieldBox/index.ts
@@ -1,3 +1,2 @@
 export * from './FieldBox';
-export * from './Label';
-export * from './BottomAddon';
+export * from './components';
